chore: remove redundant objects and further compartmentalize
